requirements.txt
typer
streamlit
numpy
torch
open-clip-torch
chromadb
scikit-learn
opencv-python
Pillow
scipy
pydantic
pyyaml
transformers
ultralytics
hf_xet